requirements.txt
django
django-ninja
django-redis